chore: Extract SwatchButton from InputColor
