+++ shared/i18n/locales/en_US/translation.json
@@ -314,8 +314,6 @@
   "Objects": "Objects",
   "Symbols": "Symbols",
   "Flags": "Flags",
-  "Select a color": "Select a color",
-  "Loading": "Loading",
   "View only": "View only",
   "Can edit": "Can edit",
   "No access": "No access",
@@ -450,6 +448,8 @@
   "Installation": "Installation",
   "Unstar document": "Unstar document",
   "Star document": "Star document",
+  "Select a color": "Select a color",
+  "Loading": "Loading",
   "Template created, go ahead and customize it": "Template created, go ahead and customize it",
   "Creating a template from <em>{{titleWithDefault}}</em> is a non-destructive action – we'll make a copy of the document and turn it into a template that can be used as a starting point for new documents.": "Creating a template from <em>{{titleWithDefault}}</em> is a non-destructive action – we'll make a copy of the document and turn it into a template that can be used as a starting point for new documents.",
   "Enable other members to use the template immediately": "Enable other members to use the template immediately",
